+++ src/services/Migrations.php
@@ -3,6 +3,7 @@ namespace verbb\navigation\services;
 
 use verbb\navigation\helpers\PluginMigrationHelper;
 use verbb\navigation\migrations\plugins\MigrateFromFreeNav;
+use verbb\navigation\migrations\plugins\MigrateFromMenuBuilder;
 use verbb\navigation\migrations\plugins\MigrateFromNavigate;
 use verbb\navigation\migrations\plugins\MigrateFromNavkit;
 use verbb\navigation\migrations\plugins\MigrateFromOlivemenus;
@@ -19,6 +20,7 @@ class Migrations extends Component
     {
         return [
             'free-nav' => $this->_sourceInfo(MigrateFromFreeNav::class, 'free-nav'),
+            'menu-builder' => $this->_sourceInfo(MigrateFromMenuBuilder::class, 'menu-builder'),
             'navigate' => $this->_sourceInfo(MigrateFromNavigate::class, 'navigate'),
             'navkit' => $this->_sourceInfo(MigrateFromNavkit::class, 'navkit'),
             'olivemenus' => $this->_sourceInfo(MigrateFromOlivemenus::class, 'olivemenus'),
@@ -30,6 +32,7 @@ class Migrations extends Component
     {
         return match ($sourceId) {
             'free-nav' => MigrateFromFreeNav::class,
+            'menu-builder' => MigrateFromMenuBuilder::class,
             'navigate' => MigrateFromNavigate::class,
             'navkit' => MigrateFromNavkit::class,
             'olivemenus' => MigrateFromOlivemenus::class,